Move source of dataset operations to file.
