+++ core/dataset_operations.cpp
@@ -1,0 +1,220 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
+/// @file
+/// @author Simon Heybrock
+#include "scipp/core/dataset.h"
+#include "scipp/core/except.h"
+#include "scipp/core/transform.h"
+
+namespace scipp::core {
+
+// For now this implementation is only for the simplest case of 2 dims (inner
+// stands for sparse)
+DataArray histogram(const DataConstProxy &sparse,
+                    const VariableConstProxy &binEdges) {
+  if (sparse.hasData())
+    throw except::SparseDataError(
+        "`histogram` is not implemented for sparse data with values yet.");
+  if (sparse.dims().ndims() > 1)
+    throw std::logic_error("Only the simple case histograms may be constructed "
+                           "for now: 2 dims including sparse.");
+  auto dim = binEdges.dims().inner();
+  if (binEdges.unit() != sparse.coords()[dim].unit())
+    throw std::logic_error(
+        "Bin edges must have same unit as the sparse input coordinate.");
+  if (binEdges.dtype() != dtype<double> ||
+      sparse.coords()[dim].dtype() != DType::Double)
+    throw std::logic_error("Histogram is only available for double type.");
+  auto coord = sparse.coords()[dim];
+  auto edgesSpan = binEdges.values<double>();
+  if (!std::is_sorted(edgesSpan.begin(), edgesSpan.end()))
+    throw std::logic_error("Bin edges should be sorted to make the histogram.");
+  auto resDims{sparse.dims()};
+  auto len = binEdges.dims()[dim] - 1;
+  resDims.resize(resDims.index(dim), len);
+  Variable result = makeVariableWithVariances<double>(resDims, units::counts);
+  for (scipp::index i = 0; i < sparse.dims().volume(); ++i) {
+    const auto &coord_i = coord.sparseValues<double>()[i];
+    auto curRes = result.values<double>().begin() + i * len;
+    for (const auto &c : coord_i) {
+      auto it = std::upper_bound(edgesSpan.begin(), edgesSpan.end(), c);
+      if (it != edgesSpan.end() && it != edgesSpan.begin())
+        ++(*(curRes + (--it - edgesSpan.begin())));
+    }
+  }
+  std::copy(result.values<double>().begin(), result.values<double>().end(),
+            result.variances<double>().begin());
+
+  std::map<Dim, Variable> coords;
+  coords.emplace(dim, binEdges);
+  return {std::move(result), std::move(coords),
+          std::map<std::string, Variable>()};
+}
+
+DataArray histogram(const DataConstProxy &sparse, const Variable &binEdges) {
+  return histogram(sparse, VariableConstProxy(binEdges));
+}
+
+Dataset histogram(const Dataset &dataset, const VariableConstProxy &bins) {
+  auto out(Dataset(DatasetConstProxy::makeProxyWithEmptyIndexes(dataset)));
+  out.setCoord(bins.dims().inner(), bins);
+  for (const auto & [ name, item ] : dataset) {
+    if (item.dims().sparse())
+      out.setData(std::string(name), histogram(item, bins));
+  }
+  return out;
+}
+
+Dataset histogram(const Dataset &dataset, const Variable &bins) {
+  return histogram(dataset, VariableConstProxy(bins));
+}
+
+Dataset histogram(const Dataset &dataset, const Dim &dim) {
+  auto bins = dataset.coords()[dim];
+  return histogram(dataset, bins);
+}
+
+Dataset merge(const DatasetConstProxy &a, const DatasetConstProxy &b) {
+  return Dataset(union_(a, b), union_(a.coords(), b.coords()),
+                 union_(a.labels(), b.labels()), union_(a.attrs(), b.attrs()));
+}
+
+/// Concatenate a and b, assuming that a and b contain bin edges.
+///
+/// Checks that the last edges in `a` match the first edges in `b`. The
+/// Concatenates the input edges, removing duplicate bin edges.
+Variable join_edges(const VariableConstProxy &a, const VariableConstProxy &b,
+                    const Dim dim) {
+  expect::equals(a.slice({dim, a.dims()[dim] - 1}), b.slice({dim, 0}));
+  return concatenate(a.slice({dim, 0, a.dims()[dim] - 1}), b, dim);
+}
+
+/// Return the dimension for given coord or labels.
+///
+/// For coords, this is the same as the key, for labels we adopt the convention
+/// that labels are "labelling" their inner dimension.
+template <class T, class Key>
+Dim dim_of_coord_or_labels(const T &dict, const Key &key) {
+  if constexpr (std::is_same_v<Key, Dim>)
+    return key;
+  else
+    return dict[key].dims().inner();
+}
+
+namespace {
+template <class T1, class T2>
+auto concat(const T1 &a, const T2 &b, const Dim dim, const Dimensions &dimsA,
+            const Dimensions &dimsB) {
+  std::map<typename T1::key_type, typename T1::mapped_type> out;
+  for (const auto & [ key, a_ ] : a) {
+    if (dim_of_coord_or_labels(a, key) == dim) {
+      if ((a_.dims()[dim] == dimsA[dim]) != (b[key].dims()[dim] == dimsB[dim]))
+        throw except::BinEdgeError(
+            "Either both or neither of the inputs must be bin edges.");
+      if (a_.dims()[dim] == dimsA[dim])
+        out.emplace(key, concatenate(a_, b[key], dim));
+      else
+        out.emplace(key, join_edges(a_, b[key], dim));
+    } else {
+      // 1D coord is kept only if both inputs have matching 1D coords.
+      if (a_.dims().contains(dim) || b[key].dims().contains(dim) ||
+          a_ != b[key])
+        out.emplace(key, concatenate(a_, b[key], dim));
+      else
+        out.emplace(key, same(a_, b[key]));
+    }
+  }
+  return out;
+}
+} // namespace
+
+DataArray concatenate(const DataConstProxy &a, const DataConstProxy &b,
+                      const Dim dim) {
+  if (!a.dims().contains(dim) && a == b)
+    return DataArray{a};
+  return DataArray(concatenate(a.data(), b.data(), dim),
+                   concat(a.coords(), b.coords(), dim, a.dims(), b.dims()),
+                   concat(a.labels(), b.labels(), dim, a.dims(), b.dims()));
+}
+
+Dataset concatenate(const DatasetConstProxy &a, const DatasetConstProxy &b,
+                    const Dim dim) {
+  Dataset result;
+  for (const auto & [ name, item ] : a)
+    if (b.contains(name))
+      result.setData(name, concatenate(item, b[name], dim));
+  return result;
+}
+
+template <class Func, class... Args>
+DataArray apply_and_drop_dim(const DataConstProxy &a, Func func, const Dim dim,
+                             Args &&... args) {
+  std::map<Dim, Variable> coords;
+  for (auto && [ d, coord ] : a.coords())
+    if (d != dim)
+      coords.emplace(d, coord);
+
+  std::map<std::string, Variable> labels;
+  for (auto && [ name, label ] : a.labels())
+    if (label.dims().inner() != dim)
+      labels.emplace(name, label);
+
+  std::map<std::string, Variable> attrs;
+  for (auto && [ name, attr ] : a.attrs())
+    if (attr.dims().inner() != dim)
+      attrs.emplace(name, attr);
+
+  return DataArray(func(a.data(), dim, std::forward<Args>(args)...),
+                   std::move(coords), std::move(labels), std::move(attrs));
+}
+
+template <class Func, class... Args>
+Dataset apply_to_items(const DatasetConstProxy &d, Func func, Args &&... args) {
+  Dataset result;
+  for (const auto & [ name, data ] : d)
+    result.setData(name, func(data, std::forward<Args>(args)...));
+  return result;
+}
+
+DataArray sum(const DataConstProxy &a, const Dim dim) {
+  return apply_and_drop_dim(a, [](auto &&... _) { return sum(_...); }, dim);
+}
+
+Dataset sum(const DatasetConstProxy &d, const Dim dim) {
+  // Currently not supporting sum/mean of dataset if one or more items do not
+  // depend on the input dimension. The definition is ambiguous (return
+  // unchanged, vs. compute sum of broadcast) so it is better to avoid this for
+  // now.
+  return apply_to_items(d, [](auto &&... _) { return sum(_...); }, dim);
+}
+
+DataArray mean(const DataConstProxy &a, const Dim dim) {
+  return apply_and_drop_dim(a, [](auto &&... _) { return mean(_...); }, dim);
+}
+
+Dataset mean(const DatasetConstProxy &d, const Dim dim) {
+  return apply_to_items(d, [](auto &&... _) { return mean(_...); }, dim);
+}
+
+DataArray rebin(const DataConstProxy &a, const Dim dim,
+                const VariableConstProxy &coord) {
+  auto rebinned = apply_and_drop_dim(
+      a, [](auto &&... _) { return rebin(_...); }, dim, a.coords()[dim], coord);
+  rebinned.setCoord(dim, coord);
+  return rebinned;
+}
+
+Dataset rebin(const DatasetConstProxy &d, const Dim dim,
+              const VariableConstProxy &coord) {
+  return apply_to_items(d, [](auto &&... _) { return rebin(_...); }, dim,
+                        coord);
+}
+
+/// Return one of the inputs if they are the same, throw otherwise.
+VariableConstProxy same(const VariableConstProxy &a,
+                        const VariableConstProxy &b) {
+  expect::equals(a, b);
+  return a;
+}
+
+} // namespace scipp::core
